% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruv_ash.R
\name{additive_fix_wrapper}
\alias{additive_fix_wrapper}
\title{Wrapper for additive_update_zl1 and additive_update_l2, mostly to
use SQUAREM.}
\usage{
additive_fix_wrapper(zlambda, Y, alpha, sig_diag)
}
\arguments{
\item{zlambda}{A vector of numerics. Z occupies all but the last
two positiions. lambda1 is the second to last position. lambda2
is the last position.}

\item{Y}{A matrix of numerics of one column. The data.}

\item{alpha}{A matrix of numerics. The estimated coefficients of
the confounders.}

\item{sig_diag}{A vector of positive numerics. The estimates of the
variances.}
}
\description{
Wrapper for additive_update_zl1 and additive_update_l2, mostly to
use SQUAREM.
}

