% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruv_ash.R
\name{additive_update_l2}
\alias{additive_update_l2}
\title{Update for lambda2 given Z and lambda in additive variance
inflation iterations.}
\usage{
additive_update_l2(Z, lambda1, lambda2, Y, alpha, sig_diag)
}
\arguments{
\item{Z}{The current estimates of the confounders}

\item{lambda1}{A scalar. The current estimate of the multiplicative
variance inflation parameter.}

\item{lambda2}{A scalar. The current estimate of the additive
variance inflation parameter.}

\item{Y}{A matrix of numerics of one column. The data.}

\item{alpha}{A matrix of numerics. The estimated coefficients of
the confounders.}

\item{sig_diag}{A vector of positive numerics. The estimates of the
variances.}
}
\description{
Update for lambda2 given Z and lambda in additive variance
inflation iterations.
}
\author{
David Gerard
}

