% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashutility.R
\name{as.data.frame.ash}
\alias{as.data.frame.ash}
\title{Create data from from ash object}
\usage{
\method{as.data.frame}{ash}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{the fitted ash object}

\item{row.names}{NULL or a character vector giving the row names for the data frame. Missing values are not allowed.}

\item{optional}{not used, included for consistency as an S3 generic/method}

\item{...}{not used, included for consistency as an S3
generic/method.}
}
\description{
Creates data frame for easy plotting of results etc
}
\details{
Returns a data frame with named columnns
}
\examples{
a = ash(rnorm(100,0,2),1)
head(as.data.frame(a))

}

