% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashn.R
\name{ash.estmode}
\alias{ash.estmode}
\title{a wrapper function that estimates the mode, using optim
called by ash if mode="estimate"}
\usage{
ash.estmode(betahat, ...)
}
\description{
a wrapper function that estimates the mode, using optim
called by ash if mode="estimate"
}

