% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mad_ash.R
\name{ash_ruv_mad}
\alias{ash_ruv_mad}
\title{Wrapper for RUV4 to ASH pipeline with MAD inflation.}
\usage{
ash_ruv_mad(Y, X, ctl, ruv_type = c("ruv4", "ruv2"), k = NULL,
  cov_of_interest = ncol(X), likelihood = c("normal", "t"),
  ash_args = list(), include_intercept = TRUE, limmashrink = FALSE)
}
\arguments{
\item{Y}{A matrix of numerics. These are the response variables
where each column has its own variance. In a gene expression
study, the rows are the individuals and the columns are the
genes.}

\item{X}{A matrix of numerics. The covariates of interest.}

\item{ctl}{A vector of logicals of length \code{ncol(Y)}. If
position i is \code{TRUE} then position i is considered a
negative control. If \code{ctl = NULL} (the default) then ASH
will be run on the OLS estimates and corresponding standard
errors.}

\item{ruv_type}{Should we run RUV4 (\code{"ruv4"}) or RUV2
(\code{"ruv2"})?}

\item{k}{A non-negative integer.The number of unobserved
confounders. If not specified and the R package sva is
installed, then this function will estimate the number of
hidden confounders using the methods of Buja and Eyuboglu
(1992).}

\item{cov_of_interest}{A positive integer. The column number of the
covariate in X whose coefficients you want to apply ASH to.}

\item{likelihood}{Either \code{"normal"} or \code{"t"}. If
\code{likelihood = "t"}, then the user may provide the degrees
of freedom by including a \code{df} element in
\code{ash_args}. If \code{ash_args$df} is \code{NULL} then the
degrees of freedom will be the sample size minus the number of
covariates minus \code{k}.}

\item{ash_args}{A list of arguments to pass to ash. See
\code{\link{ash.workhorse}} for details.}

\item{include_intercept}{A logical. If \code{TRUE}, then it will
check \code{X} to see if it has an intercept term. If not, then
it will add an intercept term. If \code{FALSE}, then \code{X}
will be unchanged.}

\item{limmashrink}{A logical. Should we apply hierarchical
shrinkage to the variances (\code{TRUE}) or not (\code{FALSE})?}
}
\description{
Wrapper for RUV4 to ASH pipeline with MAD inflation.
}

