% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ashci}
\alias{ashci}
\title{Credible Interval Computation for the ash object}
\usage{
ashci(a, level = 0.95, betaindex, lfsrcriteria = 0.05, tol = 1e-05,
  trace = FALSE, ncores = FALSE)
}
\arguments{
\item{a}{the fitted ash object}

\item{betaindex}{a vector consisting of locations of betahat where you would like to compute the credible interval}

\item{lfsrcriteria}{a scalar, in which the function would autoselect betahats based on lfsr value smaller than lfsrcriteria when index is not supplied. Setting it to 1 would compute credible interval for all observations.}

\item{tol}{the desired accuracy, default value is 1e-5.}

\item{trace}{a logical variable denoting whether some of the intermediate results of iterations should be displayed to the user. Default is FALSE.}

\item{ncores}{Whether to use parallel computing, defaults to FALSE, user could specify number of cores they would like to use. Further, if user does not specify and we have over 1000 entries to compute,  then the function would perform parallel computation  using number of CPU cores on the current host.}

\item{levels}{the level for the credible interval, (default=0.95)}
}
\value{
A matrix, with first column being the location,second column being lfsr, 3rd column being posterior mean,4th and 5th column being the lower bound and upper bound for the credible interval.
}
\description{
Given the ash object return by the main function ash, this function computes the corresponding credible interval of the mixture model.
}
\details{
Uses default optimization function and perform component-wise credible interval computation. The computation cost is linear of the length of betahat.
}
\examples{
beta = c(rep(0,20),rnorm(20))
sebetahat = abs(rnorm(40,0,1))
betahat = rnorm(40,beta,sebetahat)
beta.ash = ash(betahat, sebetahat)

CImatrix=ashci(beta.ash,level=0.95)

CImatrix1=ashci(beta.ash,level=0.95,betaindex=c(1,2,5))
CImatrix2=ashci(beta.ash,level=0.95,lfsrcriteria=0.1)
CImatrix3=ashci(beta.ash,level=0.95, betaindex=c(1:length(beta)),ncores=4)
}

