% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autumn.R
\name{ashR.autumn.demo}
\alias{ashR.autumn.demo}
\title{Print a demo plot for autumnal color palette.}
\usage{
ashR.autumn.demo(palette_name)
}
\arguments{
\item{palette_name}{The short ID for the palette, e.g. "1.8".}
}
\value{
Single plot or list of all plots as ggplot2 objects.
}
\description{
Print a demo plot for autumnal color palette.
}
\examples{
# Print demo plot for single autumn palette.
ashR.autumn("5.5")

# Print demo plots for all autumn palettes.
ashR.autumn("all")

}
