% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{ashR.autumn}
\alias{ashR.autumn}
\title{Generate autumnal color palette.}
\usage{
ashR.autumn(palette_name)
}
\arguments{
\item{palette_name}{The short ID for the palette, e.g. "1.8".}
}
\value{
A vector of hexademicals of length 5 or 8.
}
\description{
Generate autumnal color palette.
}
\examples{
# Assign palette to a name.
my_palette <- ashR.autumn("5.5")

# Concatenate two palettes for a larger custom palette.
custom <- c(ashR.autumn("2.8"), ashR.autumn("3.8"))

# Use with base R.
plot(iris$Sepal.Width,
     iris$Sepal.Length,
     col = ashR.autumn("1.5"))

}
