% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berry.R
\name{ashR.berry.demo}
\alias{ashR.berry.demo}
\title{Print a demo plot for berry-tone color palette.}
\usage{
ashR.berry.demo(palette_name)
}
\arguments{
\item{palette_name}{The short ID for the palette, e.g. "1.8".}
}
\value{
Void.
}
\description{
Print a demo plot for berry-tone color palette.
}
\examples{
# Print demo plot for single berry palette.
ashR.berry("5.5")

# Print demo plots for all berry palettes.
ashR.berry("all")

}
