% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berry.R
\name{ashR.berry}
\alias{ashR.berry}
\title{Generate berry-tone color palette.}
\usage{
ashR.berry(palette_name)
}
\arguments{
\item{palette_name}{The short ID for the palette, e.g. "1.8".}
}
\value{
A vector of hexademicals of length 3, 5, or 8.
}
\description{
Generate berry-tone color palette.
}
\examples{
# Assign palette to a name.
my_palette <- ashR.berry("5.5")

# Concatenate two palettes for a larger custom palette.
custom <- c(ashR.berry("2.8"), ashR.berry("3.8"))

# Use with base R.
plot(iris$Sepal.Width,
     iris$Sepal.Length,
     col = ashR.berry("1.5"))

}
