% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earth.R
\name{ashR.earth.demo}
\alias{ashR.earth.demo}
\title{Print a demo plot for earth-tone color palette.}
\usage{
ashR.earth.demo(palette_name)
}
\arguments{
\item{palette_name}{The short ID for the palette, e.g. "1.8".}
}
\value{
Single plot or list of all plots as ggplot2 objects.
}
\description{
Print a demo plot for earth-tone color palette.
}
\examples{
# Print demo plot for single earth palette.
ashR.earth("5.5")

# Print demo plots for all earth palettes.
ashR.earth("all")

}
