% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earth.R
\name{ashR.earth}
\alias{ashR.earth}
\title{Generate earth-tone color palette.}
\usage{
ashR.earth(palette_name)
}
\arguments{
\item{palette_name}{The short ID for the palette, e.g. "1.8".}
}
\value{
A vector of hexademicals of length 3, 5, or 8.
}
\description{
Generate earth-tone color palette.
}
\examples{
# Assign palette to a name.
my_palette <- ashR.earth("5.5")

# Concatenate two palettes for a larger custom palette.
custom <- c(ashR.earth("2.8"), ashR.earth("3.8"))

# Use with base R.
plot(iris$Sepal.Width,
     iris$Sepal.Length,
     col = ashR.earth("1.5"))

}
