% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sea.R
\name{ashR.sea.demo}
\alias{ashR.sea.demo}
\title{Print a demo plot for watery-tone color palette.}
\usage{
ashR.sea.demo(palette_name)
}
\arguments{
\item{palette_name}{The short ID for the palette, e.g. "1.8".}
}
\value{
Void.
}
\description{
Print a demo plot for watery-tone color palette.
}
\examples{
# Print demo plot for single sea palette.
ashR.sea("5.5")

# Print demo plots for all sea palettes.
ashR.sea("all")

}
