% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sea.R
\name{ashR.sea}
\alias{ashR.sea}
\title{Generate watery color palette.}
\usage{
ashR.sea(palette_name)
}
\arguments{
\item{palette_name}{The short ID for the palette, e.g. "1.8".}
}
\value{
A vector of hexademicals of length 3, 5, or 8.
}
\description{
Generate watery color palette.
}
\examples{
# Assign palette to a name.
my_palette <- ashR.sea("5.5")

# Concatenate two palettes for a larger custom palette.
custom <- c(ashR.sea("2.8"), ashR.sea("3.8"))

# Use with base R.
plot(iris$Sepal.Width,
     iris$Sepal.Length,
     col = ashR.sea("1.5"))

}
