% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calc_loglik}
\alias{calc_loglik}
\title{Compute loglikelihood for data from ash fit}
\usage{
calc_loglik(a, betahat, betahatsd, df, model = c("EE", "ES"))
}
\arguments{
\item{a}{the fitted ash object}

\item{betahat}{the data}

\item{betahatsd}{the observed standard errors}

\item{model:}{indicates whether you want the likelihood under the EE or ES model}
}
\description{
Return the log-likelihood of the data betahat, with standard errors betahatsd,
under the fitted distribution in the ash object.
}
\details{
See example in CompareBetahatvsZscoreAnalysis.rmd
}

