% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashutility.R
\name{calc_logLR}
\alias{calc_logLR}
\title{Compute loglikelihood ratio for data from ash fit}
\usage{
calc_logLR(g, data)
}
\arguments{
\item{g}{the fitted g, or an ash object containing g}

\item{data}{a data object, see set_data}
}
\description{
Return the log-likelihood ratio of the data for a given g() prior
}
