% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{calc_mixmean}
\alias{calc_mixmean}
\title{Generic function of calculating the overall mean of the mixture}
\usage{
calc_mixmean(m)
}
\arguments{
\item{m}{a mixture of k components generated by normalmix() or
unimix() or igmix()}
}
\value{
it returns scalar, the mean of the mixture distribution.
}
\description{
Generic function of calculating the overall mean of the mixture
}
