% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calc_mixsd}
\alias{calc_mixsd}
\title{Generic function of calculating the overall standard deviation of the mixture}
\usage{
calc_mixsd(m)
}
\arguments{
\item{m}{a mixture of k components generated by normalmix() or unimix() or igmix()}
}
\value{
it returns scalar
}
\description{
Generic function of calculating the overall standard deviation of the mixture
}

