% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashutility.R
\name{calc_null_vloglik}
\alias{calc_null_vloglik}
\title{Compute vector of loglikelihood for data under null that all
    beta are 0}
\usage{
calc_null_vloglik(data)
}
\arguments{
\item{data}{a data object; see set_data}
}
\description{
Return the vector of log-likelihoods of the data points under the null
}
