% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashutility.R
\name{calc_vloglik}
\alias{calc_vloglik}
\title{Compute vector of loglikelihood for data from ash fit}
\usage{
calc_vloglik(g, data)
}
\arguments{
\item{g}{the fitted g, or an ash object containing g}

\item{data}{a data object, see set_data}
}
\description{
Return the vector of log-likelihoods of the data
    betahat, with standard errors betahatsd, for a given g() prior
    on beta, or an ash object containing that
}
