% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashutility.R
\name{cdf.ash}
\alias{cdf.ash}
\title{cdf method for ash object}
\usage{
cdf.ash(a, x, lower.tail = TRUE)
}
\arguments{
\item{a}{the fitted ash object}

\item{x}{the vector of locations at which cdf is to be computed}

\item{lower.tail}{(default=TRUE) whether to compute the lower or upper tail}
}
\description{
Computed the cdf of the underlying fitted distribution
}
\details{
None
}
