% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{cdf_conv}
\alias{cdf_conv}
\title{cdf_conv}
\usage{
cdf_conv(m, data)
}
\arguments{
\item{m}{mixture distribution with k components}

\item{data}{details depend on the model}
}
\description{
compute cdf of mixture m convoluted with error distribution
 either normal of sd (s) or student t with df v at locations x
}
