% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{cdf_post}
\alias{cdf_post}
\title{cdf_post}
\usage{
cdf_post(m, c, data)
}
\arguments{
\item{m}{mixture distribution with k components}

\item{c}{a scalar}

\item{data}{details depend on model}
}
\value{
an n vector containing the cdf for beta_i at c
}
\description{
evaluate cdf of posterior distribution of beta at c. m
    is the prior on beta, a mixture; c is location of evaluation
    assumption is betahat | beta ~ t_v(beta,sebetahat)
}
\examples{
beta = rnorm(100,0,1)
betahat= beta+rnorm(100,0,1)
sebetahat=rep(1,100)
ash.beta = ash(betahat,1,mixcompdist="normal")
cdf0 = cdf_post(ash.beta$fitted_g,0,set_data(betahat,sebetahat))
graphics::plot(cdf0,1-get_pp(ash.beta))
}
