% Generated by roxygen2 (4.0.2): do not edit by hand
\name{comp_cdf}
\alias{comp_cdf}
\title{Generic function of computing the cdf for each component}
\usage{
comp_cdf(x, y, lower.tail = TRUE)
}
\arguments{
\item{x}{a mixture (eg of type normalmix or unimix)}

\item{y}{locations at which cdf to be computed}

\item{lower.tail}{boolean indicating whether to report lower tail}
}
\value{
it returns a vector of probabilities, with length equals to number of components in x
}
\description{
Generic function of computing the cdf for each component
}

