% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{comp_cdf_conv}
\alias{comp_cdf_conv}
\title{comp_cdf_conv}
\usage{
comp_cdf_conv(m, data)
}
\arguments{
\item{m}{mixture distribution with k components}

\item{data}{details depend on the model}
}
\value{
a k by n matrix of cdfs
}
\description{
compute the cdf of data for each component of mixture when convolved with error distribution
}
