% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unimix.R
\name{comp_cdf_conv.unimix}
\alias{comp_cdf_conv.unimix}
\title{cdf of convolution of each component of a unif mixture}
\usage{
\method{comp_cdf_conv}{unimix}(m, data)
}
\arguments{
\item{m}{a mixture of class unimix}

\item{data, }{see set_data()}
}
\value{
a k by n matrix
}
\description{
cdf of convolution of each component of a unif mixture
}
