% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{comp_cdf_post}
\alias{comp_cdf_post}
\title{comp_cdf_post}
\usage{
comp_cdf_post(m, c, data)
}
\arguments{
\item{m}{mixture distribution with k components}

\item{c}{a scalar}

\item{data}{details depend on model}
}
\value{
a k by n matrix
}
\description{
evaluate cdf of posterior distribution of beta at c. m
    is the prior on beta, a mixture; c is location of evaluation
    assumption is betahat | beta ~ t_v(beta,sebetahat)
}
\examples{
beta = rnorm(100,0,1)
betahat= beta+rnorm(100,0,1)
sebetahat=rep(1,100)
ash.beta = ash(betahat,1,mixcompdist="normal")
comp_cdf_post(get_fitted_g(ash.beta),0,data=set_data(beta,sebetahat))
}
