% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixlik.R
\name{comp_cdf_post_mixlik}
\alias{comp_cdf_post_mixlik}
\title{comp_cdf_post_mixlik}
\usage{
comp_cdf_post_mixlik(m, c, betahat, sebetahat, v, pilik)
}
\arguments{
\item{m}{prior on beta, a mixture}

\item{c}{location of evaluation}

\item{betahat}{the data}

\item{sebetahat}{the observed standard errors}

\item{v}{degree of freedom of error distribution}

\item{pilik}{mixture proportion}
}
\value{
it returns a (k*l) by n matrix
}
\description{
evaluate cdf of posterior distribution of beta at c
}

