% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalmix.R
\name{comp_dens_conv.normalmix}
\alias{comp_dens_conv.normalmix}
\title{comp_dens_conv.normalmix}
\usage{
\method{comp_dens_conv}{normalmix}(m, data)
}
\arguments{
\item{m}{mixture distribution with k components}

\item{data}{a list with components x and s to be interpreted as a normally-distributed observation and its standard error}
}
\value{
a k by n matrix
}
\description{
returns density of convolution of each component of a
    normal mixture with N(0,s^2) at x. Note that
    convolution of two normals is normal, so it works that way
}
