% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalmix.R
\name{comp_mean.normalmix}
\alias{comp_mean.normalmix}
\title{comp_mean.normalmix}
\usage{
\method{comp_mean}{normalmix}(m)
}
\arguments{
\item{m}{a normal mixture distribution with k components}
}
\value{
a vector of length k
}
\description{
returns mean of the normal mixture
}
