% Generated by roxygen2 (4.0.1): do not edit by hand
\name{comp_mean}
\alias{comp_mean}
\title{Generic function of calculating the first moment of components of the mixture}
\usage{
comp_mean(m)
}
\arguments{
\item{m}{a mixture of k components generated by normalmix() or unimix() or igmix()}
}
\value{
it returns a vector of means.
}
\description{
Generic function of calculating the first moment of components of the mixture
}

