% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnormalmix.R
\name{comp_mean.tnormalmix}
\alias{comp_mean.tnormalmix}
\title{comp_mean.tnormalmix}
\usage{
\method{comp_mean}{tnormalmix}(m)
}
\arguments{
\item{m}{A truncated normal mixture distribution with k components.}
}
\value{
A vector of length k.
}
\description{
Returns mean of the truncated-normal mixture.
}
