% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{comp_mean2}
\alias{comp_mean2}
\title{Generic function of calculating the second moment of components of
    the mixture}
\usage{
comp_mean2(m)
}
\arguments{
\item{m}{a mixture of k components generated by normalmix() or
unimix() or igmix()}
}
\value{
it returns a vector of second moments.
}
\description{
Generic function of calculating the second moment of components of
    the mixture
}
