% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mix.R
\name{comp_postmean.normalmix}
\alias{comp_postmean.normalmix}
\title{comp_postmean.normalmix}
\usage{
\method{comp_postmean}{normalmix}(m, betahat, sebetahat, v)
}
\arguments{
\item{m}{mixture distribution with k components}

\item{betahat}{an n vector}

\item{sebetahat}{an n vector}

\item{v}{degree of freedom of error distribution}
}
\value{
a k by n matrix
}
\description{
returns posterior mean for each component of prior m, given observations betahat and sebetahat
}

