% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{comp_postmean2}
\alias{comp_postmean2}
\title{comp_postmean2}
\usage{
comp_postmean2(m, betahat, sebetahat, v)
}
\arguments{
\item{m}{mixture distribution with k components}

\item{betahat}{an n vector of observations}

\item{sebetahat}{an n vector of standard errors}

\item{v}{degree of freedom of error distribution}
}
\description{
output posterior mean-squared value for beta for prior
    mixture m,given observations betahat, sebetahat, df v
}

