% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{comp_postmean2.unimix}
\alias{comp_postmean2.unimix}
\title{as for posterior mean, but compute posterior mean squared value}
\usage{
\method{comp_postmean2}{unimix}(m, betahat, sebetahat, v)
}
\description{
as for posterior mean, but compute posterior mean squared value
}

