% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mix.R
\name{comp_postmean2_mixlik}
\alias{comp_postmean2_mixlik}
\title{comp_postmean2_mixlik}
\usage{
comp_postmean2_mixlik(m, betahat, sebetahat, v, pilik)
}
\arguments{
\item{m}{mixture distribution}

\item{betahat}{the data}

\item{sebetahat}{the observed standard errors}

\item{v}{degree of freedom of error distribution}

\item{pilik}{mixture proportion}
}
\description{
output posterior mean-squared value for beta
}

