% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mengyin.R
\name{comp_postmean_gen}
\alias{comp_postmean_gen}
\title{comp_postmean}
\usage{
comp_postmean_gen(m, betahat, cdfFUN, pdfFUN, FUNargs, etruncFUN)
}
\description{
output posterior mean for beta for each component of
    prior mixture m, given observations betahat and FUNargs
}

