% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mix.R
\name{comp_postsd.igmix}
\alias{comp_postsd.igmix}
\title{comp_postsd.igmix}
\usage{
\method{comp_postsd}{igmix}(m, betahat, sebetahat, v)
}
\arguments{
\item{m}{a mixture with k components}

\item{betahat}{an n vector}

\item{sebetahat}{an n vector}

\item{v}{degree of freedom of error distribution}
}
\value{
a k by n matrix
}
\description{
return posterior standard deviation for each component of prior m, given observations betahat and sebetahat
}

