% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{comp_postsd}
\alias{comp_postsd}
\title{comp_postsd}
\usage{
comp_postsd(m, data)
}
\arguments{
\item{m}{mixture distribution with k components}

\item{data}{details depend on the model}
}
\description{
output posterior sd for beta for each component of
    prior mixture m,given data
}
\examples{
beta = rnorm(100,0,1)
betahat= beta+rnorm(100,0,1)
ash.beta = ash(betahat,1,mixcompdist="normal")
data=list(x=betahat,s=rep(1,100),df=NULL)
comp_postmean(ash.beta$fitted.g,data)
comp_postsd(ash.beta$fitted.g,data)
comppostprob(ash.beta$fitted.g,data)
}

