% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{comp_postsd_mixlik}
\alias{comp_postsd_mixlik}
\title{comp_postsd_mixlik}
\usage{
comp_postsd_mixlik(m, betahat, sebetahat, v, pilik)
}
\arguments{
\item{m}{mixture distribution}

\item{betahat}{the data}

\item{sebetahat}{the observed standard errors}

\item{v}{degree of freedom of error distribution}

\item{pilik}{mixture proportion}
}
\description{
output posterior sd for beta for each component of
    prior mixture m,given observations betahat, sebetahat, df v
    from l-components mixture likelihood with mixture proportion
    pilik
}

