% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalmix.R
\name{comp_sd.normalmix}
\alias{comp_sd.normalmix}
\title{comp_sd.normalmix}
\usage{
\method{comp_sd}{normalmix}(m)
}
\arguments{
\item{m}{a normal mixture distribution with k components}
}
\value{
a vector of length k
}
\description{
returns sds of the normal mixture
}
