% Generated by roxygen2 (4.0.2): do not edit by hand
\name{comp_sd}
\alias{comp_sd}
\title{Generic function of extracting the standard deviations of components of the mixture}
\usage{
comp_sd(m)
}
\arguments{
\item{m}{a mixture of k components generated by normalmix() or unimix() or igmix()}
}
\value{
it returns a vector of standard deviaitons
}
\description{
Generic function of extracting the standard deviations of components of the mixture
}

