% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnormalmix.R
\name{comp_sd.tnormalmix}
\alias{comp_sd.tnormalmix}
\title{comp_sd.normalmix}
\usage{
\method{comp_sd}{tnormalmix}(m)
}
\arguments{
\item{m}{A truncated normal mixture distribution with k components.}
}
\value{
A vector of length k.
}
\description{
Returns standard deviations of the truncated normal mixture.
}
