% Generated by roxygen2 (4.0.2): do not edit by hand
\name{compcdf_post}
\alias{compcdf_post}
\title{compcdf_post}
\usage{
compcdf_post(m, c, betahat, sebetahat, v)
}
\arguments{
\item{c}{a scalar}

\item{betahat}{an n vector of observations}

\item{sebetahat}{an n vector of standard errors}

\item{v}{degree of freedom of error distribution}
}
\value{
a k by n matrix
}
\description{
evaluate cdf of posterior distribution of beta at c,m is the prior on beta, a mixture,c is location of evaluation assumption is betahat | beta ~ N(beta,sebetahat)
}

