% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mix.R
\name{compcdf_post_mixlik}
\alias{compcdf_post_mixlik}
\title{compcdf_post_mixlik}
\usage{
compcdf_post_mixlik(m, c, betahat, sebetahat, v, pilik)
}
\arguments{
\item{m}{prior on beta, a mixture}

\item{c}{location of evaluation}

\item{betahat}{the data}

\item{sebetahat}{the observed standard errors}

\item{v}{degree of freedom of error distribution}

\item{pilik}{mixture proportion}
}
\value{
it returns a (k*l) by n matrix
}
\description{
evaluate cdf of posterior distribution of beta at c
}

