% Generated by roxygen2 (4.0.2): do not edit by hand
\name{compdens}
\alias{compdens}
\title{Generic function of calculating the component densities of the mixture}
\usage{
compdens(x, y, log = FALSE)
}
\arguments{
\item{x}{is a k-vector of mean}

\item{y}{is an n-vector of location}

\item{log}{whether to use log-scale on densities}
}
\value{
A k by n matrix of densities
}
\description{
Generic function of calculating the component densities of the mixture
}

