% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{compdens_conv}
\alias{compdens_conv}
\title{compdens_conv}
\usage{
compdens_conv(m, data)
}
\arguments{
\item{m}{mixture distribution with k components}

\item{data}{details depend on the model}
}
\value{
a k by n matrix of densities
}
\description{
compute the density of data for each component of mixture when convolved with error distribution
}

