% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unimix.R
\name{compdens_conv.unimix}
\alias{compdens_conv.unimix}
\title{density of convolution of each component of a unif mixture}
\usage{
\method{compdens_conv}{unimix}(m, data)
}
\arguments{
\item{m}{a mixture of class unimix}

\item{data, }{see set_data()}
}
\value{
a k by n matrix
}
\description{
density of convolution of each component of a unif mixture
}

