% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{compdens_conv_mixlik}
\alias{compdens_conv_mixlik}
\title{compdens_conv_mixlik}
\usage{
compdens_conv_mixlik(m, x, s, v, pilik, FUN = "+")
}
\arguments{
\item{m}{a mixture with k components}

\item{x}{an n vector}

\item{s}{normal mixture of sd(s), n-by-l matrix}

\item{v}{degree of freedom of error distribution}

\item{pilik}{mixture proportion , n-by-l matrix}

\item{FUN}{default is "+"}
}
\value{
a (k*l) by n matrix of densities
}
\description{
compute the density of the components of the mixture m
    when convoluted with l-components normal mixture with standard
    deviation s or C (scale vector) multiplies scaled (se)
    student.t l-components mixture with df v with mixture
    proportion pilik the density is evaluated at x
}

