% Generated by roxygen2 (4.0.2): do not edit by hand
\name{comppostprob.default}
\alias{comppostprob.default}
\title{comppostprob.default}
\usage{
\method{comppostprob}{default}(m, x, s, v)
}
\arguments{
\item{m}{mixture distribution}

\item{x}{an n vector}

\item{s}{an n vector or integer}

\item{v}{degree of freedom of error distribution}
}
\description{
comppostprob.default
}

