% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{comppostprob_mixlik}
\alias{comppostprob_mixlik}
\title{comppostprob_mixlik}
\usage{
comppostprob_mixlik(m, x, s, v, pilik)
}
\arguments{
\item{m}{mixture distribution}

\item{x}{an n vector}

\item{s}{normal mixture of sd(s), n-by-l matrix}

\item{v}{degree of freedom of error distribution}

\item{pilik}{mixture proportion , n-by-l matrix}
}
\description{
compute the posterior prob that each observation came
    from each component of the mixture m,output a k by n vector of
    probabilities computed by weighting the component densities by
    pi and then normalizing,when likelihood is an l-components
    normal mixture or student t mixture with mixture proportion
    pilik
}

