% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cxxMixSquarem}
\alias{cxxMixSquarem}
\title{Brief description of function.}
\usage{
cxxMixSquarem(matrix_lik, prior, pi_init, control)
}
\arguments{
\item{matrix_lik}{Description of argument goes here.}

\item{prior}{Description of argument goes here.}

\item{pi_init}{Description of argument goes shere.}

\item{control}{Description of argument goes here.}
}
\description{
Explain here what this function does.
}
