% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{dens}
\alias{dens}
\title{Find density at y, a generic function}
\usage{
dens(x, y)
}
\arguments{
\item{x}{A mixture of k components generated by
\code{\link{normalmix}} or \code{\link{unimix}}.}

\item{y}{An n-vector of the location.}
}
\description{
Find density at y, a generic function
}
