% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mix.R
\name{dens_conv}
\alias{dens_conv}
\title{dens_conv}
\usage{
dens_conv(m, x, s, v, FUN = "+")
}
\arguments{
\item{m}{mixture distribution}

\item{x}{an n vector}

\item{s}{an n vector or integer}

\item{v}{degree of freedom of error distribution}

\item{FUN}{default is "+"}
}
\description{
compute density of mixture m convoluted with normal of sd (s) or student t with df v at locations x
}

