% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{dens_conv_mixlik}
\alias{dens_conv_mixlik}
\title{dens_conv_mixlik}
\usage{
dens_conv_mixlik(m, x, s, v, pilik, FUN = "+")
}
\arguments{
\item{m}{mixture distribution}

\item{x}{an n vector}

\item{s}{normal mixture of sd(s), n-by-l matrix}

\item{v}{degree of freedom of error distribution}

\item{pilik}{mixture proportion , n-by-l matrix}

\item{FUN}{default is "+"}
}
\description{
compute density of mixture m convoluted with
    l-components normal mixture of sd (s) or student t mixture with
    df v with mixture proportion pilik at locations x
}

