\name{density.ash}
\alias{density.ash}
\title{Density method for ash object}
\usage{
  density.ash(a, x)
}
\arguments{
  \item{a}{the fitted ash object}

  \item{x}{the vector of locations at which density is to
  be computed}
}
\description{
  Return the density of the underlying fitted distribution
}
\details{
  None
}

