% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashutility.R
\name{diag.ash}
\alias{diag.ash}
\title{Diagnostic plot for ash object}
\usage{
diag.ash(x, ...)
}
\arguments{
\item{x}{the fitted ash object}

\item{...}{Arguments to be passed to methods,such as graphical parameters (see \code{\link[graphics]{plot}})}
}
\description{
Plot the cdf of the fitted predictive distribution at observations
}
\details{
None
}
