% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logF.R
\name{dlogf}
\alias{dlogf}
\title{The log-F distribution}
\usage{
dlogf(x, df1, df2, ncp, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles}

\item{df1}{degrees of freedom}

\item{df2}{degrees of freedom}

\item{ncp}{non-centrality parameter. If omitted the central F is assumed.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
The density function.
}
\description{
Density function for the log-F distribution with \code{df1} and \code{df2}
 degrees of freedom (and optional non-centrality parameter \code{ncp}).
}
