% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashgen.R
\name{dtgen}
\alias{dtgen}
\title{Generalized student t distribution}
\usage{
dtgen(x, df, ncp, mean = 0, sd = 1, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles}

\item{df}{degrees of freedom (> 0, maybe non-integer).}

\item{ncp}{non-centrality parameter delta; only for abs(ncp) <= 37.62. 
If omitted, use the central t distribution.}

\item{mean}{location parameter}

\item{sd}{scale parameter}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
The density.
}
\description{
Density function for the generalized t distribution with 
      \code{df} degrees of freedom, location parameter \code{mean}, scale parameter \code{sd} 
      (and optional non-centrality parameter \code{ncp}). 
      i.e. (X-\code{mean})/\code{sd} follows standard t-distribution with \code{df} and \code{ncp}.
}

