% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earth.R
\docType{data}
\name{earth}
\alias{earth}
\title{List of earth-tone palettes. Palettes are 5, 3, or 8 values long.}
\format{
An object of class \code{list} of length 15.
}
\usage{
earth
}
\description{
List of earth-tone palettes. Palettes are 5, 3, or 8 values long.
}
\keyword{datasets}
