% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mengyin.R
\name{estimate_mixprop_gen}
\alias{estimate_mixprop_gen}
\title{Estimate mixture proportions of sigmaa by EM algorithm}
\usage{
estimate_mixprop_gen(betahat, cdfFUN, pdfFUN, FUNargs, g, prior,
  optmethod = c("mixEM", "mixVBEM", "cxxMixSquarem", "mixIP"),
  null.comp = 1, control = list())
}
\arguments{
\item{betahat}{a p vector of estimates}

\item{cdfFUN}{the CDF function of error distribution. 
Defaults for likelihood "normal", "t" and "logF" are "pnorm", 
"ptgen" and "plogf" respectively.}

\item{pdfFUN}{the PDF function of error distribution.
Defaults for likelihood "normal", "t" and "logF" are "dnorm", 
"dtgen" and "dlogf" respectively.}

\item{FUNargs}{a list of arguments passed to \code{cdfFUN} & \code{pdfFUN}. 
Each argument must be a scalar or a p vector corresponding to \code{betahat}.}

\item{g}{the prior distribution for beta (usually estimated from
the data; this is used primarily in simulated data to do
computations with the "true" g)}

\item{prior}{string, or numeric vector indicating Dirichlet prior
on mixture proportions (defaults to "uniform", or (1,1...,1);
also can be "nullbiased" (nullweight,1,...,1) to put more
weight on first component), or "unit" (1/K,...,1/K) [for
optmethod=mixVBEM version only]}

\item{optmethod}{specifies optimization method used. Default is
"mixIP", an interior point method, if REBayes is installed;
otherwise an EM algorithm is used. The interior point method is
faster for large problems (n>2000).}

\item{null.comp}{the position of the null component}

\item{control}{A list of control parameters for the optmization
algorithm. Default value is set to be control.default=list(K =
1, method=3, square=TRUE, step.min0=1, step.max0=1, mstep=4,
kr=1, objfn.inc=1,tol=1.e-07, maxiter=5000, trace=FALSE). User
may supply changes to this list of parameter, say,
control=list(maxiter=10000,trace=TRUE)}
}
\value{
A list, including the final loglikelihood, the null loglikelihood, 
       a n by k likelihoodmatrix with (j,k)th element equal to \eqn{f_k(x_j)},
       and a flag to indicate convergence.
}
\description{
Estimate mixture proportions of sigmaa by EM algorithm
}

