% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncgen.R
\name{gen_etruncFUN}
\alias{gen_etruncFUN}
\title{gen_etruncFUN}
\usage{
gen_etruncFUN(lcdfFUN, lpdfFUN)
}
\arguments{
\item{lcdfFUN}{the log cdfFUN of the error distribution}

\item{lpdfFUN}{the log pdfFUN of the error distribution}
}
\description{
Produce function to compute expectation of truncated 
   error distribution from log cdf and log pdf (using numerical integration)
}
