% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igmix.R
\name{igmix}
\alias{igmix}
\title{Constructor for igmix class}
\usage{
igmix(pi, alpha, beta)
}
\arguments{
\item{pi}{vector of mixture proportions}

\item{alpha}{vector of shape parameters}

\item{beta}{vector of rate parameters}
}
\value{
an object of class igmix
}
\description{
Creates an object of class igmix (finite mixture of
    univariate inverse-gammas)
}
\details{
None
}
\examples{
igmix(c(0.5,0.5),c(1,1),c(1,2))

}
