% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lik.R
\name{lik_normal}
\alias{lik_normal}
\title{Likelihood object for normal error distribution}
\usage{
lik_normal()
}
\description{
Creates a likelihood object for ash for use with normal error distribution
}
\examples{
   z = rnorm(100) + rnorm(100) # simulate some data with normal error
   ash(z,1,lik=lik_normal())
}
