% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lik.R
\name{lik_t}
\alias{lik_t}
\title{Likelihood object for t error distribution}
\usage{
lik_t(df)
}
\arguments{
\item{df}{degree of freedom parameter of t distribution}
}
\description{
Creates a likelihood object for ash for use with t error distribution
}
\examples{
   z = rnorm(100) + rt(100,df=4) # simulate some data with t error
   ash(z,1,lik=lik_t(df=4))
}
