% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unimix.R
\name{log_comp_dens_conv.unimix}
\alias{log_comp_dens_conv.unimix}
\title{log density of convolution of each component of a unif mixture}
\usage{
\method{log_comp_dens_conv}{unimix}(m, data)
}
\arguments{
\item{m}{a mixture of class unimix}

\item{data}{see set_data()}
}
\value{
a k by n matrix of densities
}
\description{
log density of convolution of each component of a unif mixture
}
