% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{log_compdens_conv}
\alias{log_compdens_conv}
\title{log_compdens_conv}
\usage{
log_compdens_conv(m, x, s, v, FUN = "+")
}
\arguments{
\item{m}{mixture distribution}

\item{x}{an n vector}

\item{s}{an n vector or integer}

\item{v}{degree of freedom of error distribution}

\item{FUN}{default is "+"}
}
\value{
a k by n matrix of log densities
}
\description{
compute the log density of the components of the
    mixture m when convoluted with a normal with standard deviation
    s or a scaled (se) student.t with df v, the density is
    evaluated at x
}

