% Generated by roxygen2 (4.0.1): do not edit by hand
\name{loglik.ash}
\alias{loglik.ash}
\title{Compute loglikelihood for data from ash fit}
\usage{
\method{loglik}{ash}(a, betahat, betahatsd, df, zscores = FALSE)
}
\arguments{
\item{a}{the fitted ash object}

\item{betahat}{the data}

\item{betahatsd}{the observed standard errors}

\item{zscores}{indicates whether ash object was originally fit to z scores but you want the likelihood for the betahats}
}
\description{
Return the log-likelihood of the data betahat, with standard errors betahatsd,
under the fitted distribution in the ash object.
}
\details{
See example in CompareBetahatvsZscoreAnalysis.rmd
}

