% Generated by roxygen2 (4.0.2): do not edit by hand
\name{loglik_conv.default}
\alias{loglik_conv.default}
\title{loglik_conv.default}
\usage{
\method{loglik_conv}{default}(m, betahat, betahatsd, v, FUN = "+")
}
\arguments{
\item{m}{mixture distribution with k components}

\item{betahat}{an n vector of observations}

\item{v}{degree of freedom of error distribution}

\item{sebetahat}{an n vector of standard errors}
}
\description{
loglik_conv.default
}

