% Generated by roxygen2 (4.0.2): do not edit by hand
\name{loglik_conv}
\alias{loglik_conv}
\title{loglik_conv}
\usage{
loglik_conv(m, betahat, betahatsd, v, FUN = "+")
}
\arguments{
\item{m}{mixture distribution with k components}

\item{betahat}{an n vector of observations}

\item{v}{degree of freedom of error distribution}

\item{sebetahat}{an n vector of standard errors}
}
\description{
find log likelihood of data in betahat, when the mixture m is convolved with a normal with sd betahatsd
}

