% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mix.R
\name{loglik_conv_mixlik.default}
\alias{loglik_conv_mixlik.default}
\title{loglik_conv.default}
\usage{
\method{loglik_conv_mixlik}{default}(m, betahat, betahatsd, v, pilik,
  FUN = "+")
}
\arguments{
\item{m}{mixture distribution}

\item{betahat}{an n vector}

\item{betahatsd}{an n-by-l matrix}

\item{v}{degree of freedom of error distribution}

\item{pilik}{an l vector}

\item{FUN}{default is "+"}
}
\description{
loglik_conv.default
}

