% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mix.R
\name{loglik_conv_mixlik}
\alias{loglik_conv_mixlik}
\title{loglik_conv_mixlik}
\usage{
loglik_conv_mixlik(m, betahat, betahatsd, v, pilik, FUN = "+")
}
\arguments{
\item{m}{mixture distribution}

\item{betahat}{an n vector}

\item{betahatsd}{an n-by-l matrix}

\item{v}{degree of freedom of error distribution}

\item{pilik}{an l vector}

\item{FUN}{default is "+"}
}
\description{
find log likelihood of data, when the mixture m is convolved with l-comp normal mixture in betahat with mixture sd betahatsd, mixture proportion pilik.
}

