% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{mixcdf.default}
\alias{mixcdf.default}
\title{mixcdf.default}
\usage{
\method{mixcdf}{default}(x, y, lower.tail = TRUE)
}
\arguments{
\item{x}{a mixture (eg of type normalmix or unimix)}

\item{y}{locations at which cdf to be computed}

\item{lower.tail}{boolean indicating whether to report lower tail}
}
\description{
The default version of \code{\link{mixcdf}}.
}
