% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{mixcdf}
\alias{mixcdf}
\title{mixcdf}
\usage{
mixcdf(x, y, lower.tail = TRUE)
}
\arguments{
\item{x}{a mixture (eg of type normalmix or unimix)}

\item{y}{locations at which cdf to be computed}

\item{lower.tail}{boolean indicating whether to report lower tail}
}
\value{
an object of class normalmix
}
\description{
Returns cdf for a mixture (generic function)
}
\details{
None
}
\examples{
mixcdf(normalmix(c(0.5,0.5),c(0,0),c(1,2)),seq(-4,4,length=100))

}
