% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{mixprop}
\alias{mixprop}
\title{Generic function of extracting the mixture proportions}
\usage{
mixprop(m)
}
\arguments{
\item{m}{a mixture of k components generated by normalmix() or
unimix() or igmix()}
}
\value{
it returns a vector of component probabilities, summing up
    to 1.
}
\description{
Generic function of extracting the mixture proportions
}
