% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_opt.R
\name{mixSQP}
\alias{mixSQP}
\title{Estimate mixture proportions of a mixture model using
mix-SQP algorithm.}
\usage{
mixSQP(matrix_lik, prior, pi_init = NULL, control = list(), weights)
}
\description{
Estimate mixture proportions of a mixture model using
mix-SQP algorithm.
}
