% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncbeta.R
\name{my_e2truncbeta}
\alias{my_e2truncbeta}
\title{second moment of truncated Beta distribution}
\usage{
my_e2truncbeta(a, b, alpha, beta)
}
\arguments{
\item{a}{left limit of distribution}

\item{b}{right limit of distribution}

\item{alpha, beta}{shape parameters of Beta distribution}
}
\description{
Compute second moment of the truncated Beta.
}
