% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncgamma.R
\name{my_e2truncgamma}
\alias{my_e2truncgamma}
\title{second moment of truncated gamma distribution}
\usage{
my_e2truncgamma(a, b, shape, rate)
}
\arguments{
\item{a}{left limit of distribution}

\item{b}{right limit of distribution}

\item{shape}{shape of gamma distribution}

\item{rate}{rate of gamma distribution}
}
\description{
Compute second moment of the truncated gamma.
}
