% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{my_e2truncnorm}
\alias{my_e2truncnorm}
\title{More about the truncated normal}
\usage{
my_e2truncnorm(a, b, mean = 0, sd = 1)
}
\arguments{
\item{a}{Left limit of distribution.}

\item{b}{Right limit of distribution.}

\item{mean}{The mean of the untruncated normal.}

\item{sd}{The standard deviation of the untruncated normal.}
}
\description{
More about the truncated normal
}

