% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{my_e2trunct}
\alias{my_e2trunct}
\title{my_etrunct}
\usage{
my_e2trunct(a, b, v, n = 2)
}
\arguments{
\item{a}{left limit of distribution}

\item{b}{right limit of distribution}

\item{v}{degree of freedom of error distribution}

\item{n}{is moment, the default is 2 which means second moment}
}
\description{
Compute second moment of the truncated t. The result is from the paper "Moments of truncated t and F distributions" by Saralees Nadarajah and Samuel Kotz.
}

