% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncbeta.R
\name{my_etruncbeta}
\alias{my_etruncbeta}
\title{mean of truncated Beta distribution}
\usage{
my_etruncbeta(a, b, alpha, beta)
}
\arguments{
\item{a}{left limit of distribution}

\item{b}{right limit of distribution}

\item{alpha, beta}{shape parameters of Beta distribution}
}
\description{
Compute mean of the truncated Beta.
}
