% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncgen.R
\name{my_etruncdistn}
\alias{my_etruncdistn}
\title{my_etruncdistn}
\usage{
my_etruncdistn(a, b, cdfFUN, pdfFUN, FUNargs, etruncFUN)
}
\arguments{
\item{a}{Left limit of distribution.}

\item{b}{Right limit of distribution.}

\item{cdfFUN}{the CDF function of error distribution. 
Defaults for likelihood "normal", "t" and "logF" are "pnorm", 
"ptgen" and "plogf" respectively.}

\item{pdfFUN}{the PDF function of error distribution.
Defaults for likelihood "normal", "t" and "logF" are "dnorm", 
"dtgen" and "dlogf" respectively.}

\item{FUNargs}{a list of arguments passed to \code{cdfFUN} & \code{pdfFUN}. 
Each argument must be a scalar or a p vector corresponding to \code{betahat}.}

\item{etruncFUN}{(optional) the function to compute expectation of the truncated 
error distribution, which will be used for calculating \code{PosteriorMean}. 
The function should take \code{(a,b,...)}
as inputs and return expectation of truncated error distribution on (\code{a,b}) 
with distribution parameters \code{...}. Default method uses numerical integration
to compute the truncated mean.}
}
\description{
Compute expectation of truncated error distribution.
}

