% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncgamma.R
\name{my_etruncgamma}
\alias{my_etruncgamma}
\title{mean of truncated gamma distribution}
\usage{
my_etruncgamma(a, b, shape, rate)
}
\arguments{
\item{a}{left limit of distribution}

\item{b}{right limit of distribution}

\item{shape}{shape of gamma distribution}

\item{rate}{rate of gamma distribution}
}
\description{
Compute mean of the truncated gamma.
}
