% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logF.R
\name{my_etrunclogf}
\alias{my_etrunclogf}
\title{my_etrunclogf}
\usage{
my_etrunclogf(a, b, df1, df2)
}
\arguments{
\item{a}{Left limit of distribution.}

\item{b}{Right limit of distribution.}

\item{df1, df2}{degrees of freedom}
}
\description{
Compute expectation of truncated log-F distribution.
}
