% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{my_etrunct}
\alias{my_etrunct}
\title{my_etrunct}
\usage{
my_etrunct(a, b, v)
}
\arguments{
\item{a}{left limit of distribution}

\item{b}{right limit of distribution}

\item{v}{degree of freedom of error distribution}
}
\description{
Compute expectation of truncated t, the result is from
    the paper
    "Moments of truncated Student-t distribution by Hea-Jung Kim"
}

