% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mix.R
\name{ncomp.default}
\alias{ncomp.default}
\title{ncomp.default}
\usage{
\method{ncomp}{default}(m)
}
\arguments{
\item{m}{a mixture of k components generated by normalmix() or unimix() or igmix()}
}
\description{
ncomp.default
}

