% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ncomp}
\alias{ncomp}
\title{ncomp}
\usage{
ncomp(m)
}
\arguments{
\item{m}{a mixture of k components generated by normalmix() or unimix() or igmix()}
}
\description{
Given the ash object return by the main function ash, this function computes the corresponding credible interval of the mixture model.
}

