% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nonzeromeanEM}
\alias{nonzeromeanEM}
\title{Estimate unimodal nonzero mean of a mixture model by EM algorithm}
\usage{
nonzeromeanEM(betahat, sebetahat, mixsd, mixcompdist, pi.init = NULL,
  tol = 1e-07, maxiter = 5000)
}
\arguments{
\item{betahat,}{a p vector of estimates}

\item{sebetahat,}{a p vector of corresponding standard errors}

\item{pi.init,}{the initial value of \eqn{\pi} to use. If not specified defaults to (1/k,...,1/k).}

\item{tol,}{the tolerance for convergence of log-likelihood.}

\item{maxiter}{the maximum number of iterations performed}

\item{mixsd:}{vector of sds for underlying mixture components}
}
\value{
A list, including the estimates (\eqn{\mu}) and (\eqn{\pi}), the log likelihood for each iteration (NQ)
and a flag to indicate convergence
}
\description{
Given the data, standard error of the data and standard deviations of the Gaussian mixture model, estimate the mean of a unimodal Gaussian mixture by an EM algorithm.
}
\details{
Fits a k component mixture model \deqn{f(x|\pi) = \sum_k \pi_k f_k(x)} to independent
and identically distributed data \eqn{x_1,\dots,x_n}.
Estimates unimodal mean \eqn{\mu} by EM algorithm. Uses the SQUAREM package to accelerate convergence of EM. Used by the ash main function; there is no need for a user to call this
function separately, but it is exported for convenience.
Currently works for mixcompdist="normal", while "uniform" abd "halfuniform" would only return the naive mean
}

