% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ash.R
\name{nonzeromodeEM}
\alias{nonzeromodeEM}
\title{Estimate unimodal nonzero mean of a mixture model by EM algorithm}
\usage{
nonzeromodeEM(betahat, sebetahat, mixsd, mixcompdist, df = NULL,
  pi.init = NULL, retol = 1e-05, maxiter = 500, trace = trace, K = 1)
}
\arguments{
\item{betahat}{a p vector of estimates.}

\item{sebetahat}{a p vector of corresponding standard errors.}

\item{mixsd}{vector of sds for underlying mixture components.}

\item{mixcompdist}{distribution of components in mixture ( "uniform","halfuniform" or "normal").}

\item{df}{appropriate degrees of freedom for (t) distribution of betahat/sebetahat, default is NULL(Gaussian).}

\item{pi.init}{the initial value of \eqn{\pi} to use. If not specified defaults to (1/k,...,1/k).}

\item{retol}{the relative tolerance for estimate of nonzero mean,default is 1e-5.}

\item{maxiter}{the maximum number of iterations performed, default is 500.}

\item{trace}{a logical variable denoting whether some of the intermediate results of iterations should be displayed to the user, default is FALSE.}

\item{K}{An integer denoting the order of the SQUAREM scheme. Default is 1,i.e. first-order schemes, which is adequate for most problems. K=2,3 may provide greater speed in some problems, although they are less reliable than the first-order schemes.}
}
\value{
A list, including the estimates (\eqn{\mu}) and (\eqn{\pi}), the log likelihood for each iteration (NQ)
and a flag to indicate convergence.
}
\description{
Given the data, standard error of the data and standard deviations of the Gaussian mixture model, estimate the mean of a unimodal Gaussian mixture by an EM algorithm.
}
\details{
Fits a k component mixture model \deqn{f(x|\pi) = \sum_k \pi_k f_k(x)} to independent
and identically distributed data \eqn{x_1,\dots,x_n}.
Estimates unimodal mean \eqn{\mu} by EM algorithm. Uses the SQUAREM package to accelerate convergence of EM. Used by the ash main function; there is no need for a user to call this
function separately, but it is exported for convenience.
}

