% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalmix.R
\name{normalmix}
\alias{normalmix}
\title{Constructor for normalmix class}
\usage{
normalmix(pi, mean, sd)
}
\arguments{
\item{pi}{vector of mixture proportions}

\item{mean}{vector of means}

\item{sd}{vector of standard deviations}
}
\value{
an object of class normalmix
}
\description{
Creates an object of class normalmix (finite mixture
    of univariate normals)
}
\details{
None
}
\examples{
normalmix(c(0.5,0.5),c(0,0),c(1,2))

}
