% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lik.R
\name{normalmix_lik}
\alias{normalmix_lik}
\title{Likelihood object for normal mixture error distribution}
\usage{
normalmix_lik(pilik, sdlik)
}
\description{
Creates a likelihood object for ash for use with normal mixture error distribution
}
\examples{
   e = rnorm(100,0,0.8) 
   e[seq(1,100,by=2)] = rnorm(50,0,1.5) # generate e~0.5*N(0,0.8^2)+0.5*N(0,1.5^2)
   betahat = rnorm(100)+e
   ash(betahat, 1, lik=normalmix_lik(c(0.5,0.5),c(0.8,1.5)))
}

