% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{old.comppostprob.default}
\alias{old.comppostprob.default}
\title{The old default version of \code{\link{comppostprob}}.}
\usage{
old.comppostprob.default(m, x, s, v)
}
\arguments{
\item{m}{mixture distribution}

\item{x}{an n vector}

\item{s}{an n vector or integer}

\item{v}{degree of freedom of error distribution}
}
\description{
The old default version of \code{\link{comppostprob}}.
}

