% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{pcdf_post}
\alias{pcdf_post}
\title{pcdf_post}
\usage{
pcdf_post(m, c, data)
}
\arguments{
\item{m}{mixture distribution with k components}

\item{c}{a numeric vector with n elements}

\item{data}{depends on context}
}
\value{
an n vector, whose ith element is the cdf for beta_i at c_i
}
\description{
``parallel" vector version of \code{\link{cdf_post}} where c is a vector, of same length as betahat and sebetahat
}
\examples{
beta = rnorm(100,0,1)
betahat= beta+rnorm(100,0,1)
sebetahat=rep(1,100)
ash.beta = ash(betahat,1,mixcompdist="normal")
c = pcdf_post(get_fitted_g(ash.beta),beta,set_data(betahat,sebetahat))
}
