% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.ash}
\alias{plot.ash}
\title{Plot method for ash object}
\usage{
\method{plot}{ash}(a, xmin, xmax, ...)
}
\arguments{
\item{a}{the fitted ash object}

\item{xmin}{xlim lower range}

\item{xmax}{xlim upper range}

\item{...}{Arguments to be passed to methods,such as graphical parameters (see \code{\link[graphics]{plot}})}
}
\description{
Plot the density of the underlying fitted distribution
}
\details{
None
}

