% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashutility.R
\name{plot_diagnostic}
\alias{plot_diagnostic}
\title{Diagnostic plot for ash object}
\usage{
plot_diagnostic(x, xlim = c(0, 1), ylim = c(0, 1),
  xlab = "Theoretical Uniform Quantile",
  ylab = "Estimated Predictive Quantile", main = "Diagnostic Plot for ASH",
  pch = 19, cex = 0.25, ...)
}
\arguments{
\item{x}{the fitted ash object}

\item{...}{Arguments to be passed to methods,such as graphical parameters (see \code{\link[graphics]{plot}})}
}
\description{
Plot the cdf of the fitted predictive distribution at observations
}
\details{
None
}
