% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{pm_on_zero}
\alias{pm_on_zero}
\title{Generic function to extract which components of mixture are point mass on 0}
\usage{
pm_on_zero(m)
}
\arguments{
\item{m}{a mixture of k components generated by normalmix() or unimix() or igmix()}
}
\value{
a boolean vector indicating which components are point mass on 0
}
\description{
Generic function to extract which components of mixture are point mass on 0
}
