% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lik.R
\name{pois_lik}
\alias{pois_lik}
\title{Likelihood object for Poisson error distribution}
\usage{
pois_lik(y, link = c("identity", "log"))
}
\arguments{
\item{y}{Poisson observations}

\item{link}{Link function. The "identity" link directly puts unimodal prior on Poisson
intensities lambda, and "log" link puts unimodal prior on log(lambda).}
}
\description{
Creates a likelihood object for ash for use with Poisson error distribution
}
\details{
Suppose we have Poisson observations \code{y} where \eqn{y_i\sim Poisson(\p_i)}. 
   We either put an unimodal prior on the intensities \eqn{p_i\sim g} (by specifying 
   \code{link="identity"}) or on the logit intensities \eqn{logit(p_i)\sim g} (by specifying
   \code{link="logit"}). Either way, ASH with this Poisson likelihood function will compute 
   the posterior mean of the intensities \eqn{p_i}.
}
\examples{
   beta = c(rnorm(100,50,5)) # prior mode: 50
   y = rpois(100,beta) # simulate Poisson observations
   ash(rep(0,length(y)),1,lik=pois_lik(y))
}

