% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalmix.R
\name{post_sample.normalmix}
\alias{post_sample.normalmix}
\title{post_sample.normalmix}
\usage{
\method{post_sample}{normalmix}(m, data, nsamp)
}
\arguments{
\item{m}{mixture distribution with k components}

\item{data}{a list with components x and s to be interpreted as a 
normally-distributed observation and its standard error}

\item{nsamp}{number of samples to return for each observation}
}
\value{
a nsamp by n matrix
}
\description{
returns random samples from the posterior, given a
  prior distribution m and n observed datapoints.
}
