% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unimix.R
\name{post_sample.unimix}
\alias{post_sample.unimix}
\title{post_sample.unimix}
\usage{
\method{post_sample}{unimix}(m, data, nsamp)
}
\arguments{
\item{m}{mixture distribution with k components}

\item{data}{a list with components x and s to be interpreted as a 
normally-distributed observation and its standard error}

\item{nsamp}{number of samples to return for each observation}
}
\value{
a nsamp by n matrix
}
\description{
returns random samples from the posterior, given a
  prior distribution m and n observed datapoints.
}
