% Generated by roxygen2 (4.0.2): do not edit by hand
\name{postmean.default}
\alias{postmean.default}
\title{postmean.default}
\usage{
\method{postmean}{default}(m, betahat, sebetahat, v)
}
\arguments{
\item{m}{mixture distribution with k components}

\item{betahat}{an n vector of observations}

\item{sebetahat}{an n vector of standard errors}

\item{v}{degree of freedom of error distribution}
}
\description{
postmean.default
}

