% Generated by roxygen2 (4.0.2): do not edit by hand
\name{postsd_mixlik}
\alias{postsd_mixlik}
\title{postsd_mixlik}
\usage{
postsd_mixlik(m, betahat, sebetahat, v, pilik)
}
\arguments{
\item{m}{mixture distribution}

\item{betahat}{the data}

\item{sebetahat}{the observed standard errors}

\item{v}{degree of freedom of error distribution}

\item{pilik}{mixture proportion}
}
\description{
output posterior sd for beta for prior mixture m
}

