% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashutility.R
\name{print.ash}
\alias{print.ash}
\title{Print method for ash object}
\usage{
\method{print}{ash}(x, ...)
}
\arguments{
\item{x}{the fitted ash object}

\item{...}{not used, included for consistency as an S3
generic/method.}
}
\description{
Print the fitted distribution of beta values in the EB
    hierarchical model
}
\details{
None
}
