% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_args.R
\name{process_args}
\alias{process_args}
\title{Process input arguments for ash.workhorse}
\usage{
process_args(oldargs)
}
\arguments{
\item{oldargs}{captured argument list}
}
\value{
list containing the processed arguments
}
\description{
Process input arguments for ash.workhorse
}

