% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mengyin.R
\name{process_args_gen}
\alias{process_args_gen}
\title{Process input arguments for ashgen}
\usage{
process_args_gen(oldargs)
}
\arguments{
\item{oldargs}{captured argument list}
}
\value{
list containing the processed arguments
}
\description{
Process input arguments for ashgen
}

