% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{prune}
\alias{prune}
\title{prune}
\usage{
prune(m, thresh = 1e-10)
}
\arguments{
\item{m}{What is this argument?}

\item{thresh}{the threshold below which components are removed}
}
\description{
prunes out mixture components with low weight
}
