% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashgen.R
\name{ptgen}
\alias{ptgen}
\title{Generalized student t distribution}
\usage{
ptgen(q, df, ncp, mean = 0, sd = 1, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{q}{vector of quantiles}

\item{df}{degrees of freedom (> 0, maybe non-integer).}

\item{ncp}{non-centrality parameter delta; only for abs(ncp) <= 37.62. 
If omitted, use the central t distribution.}

\item{mean}{location parameter}

\item{sd}{scale parameter}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X ≤ x], otherwise, P[X > x].}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
The distribution function.
}
\description{
Distribution function for the generalized t distribution with 
      \code{df} degrees of freedom, location parameter \code{mean}, scale parameter \code{sd} 
      (and optional non-centrality parameter \code{ncp}). 
      i.e. (X-\code{mean})/\code{sd} follows standard t-distribution with \code{df} and \code{ncp}.
}

