% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix_like.R
\name{qnormalmix}
\alias{qnormalmix}
\title{Quantile function for a mixture of normals.}
\usage{
qnormalmix(p, mixdense)
}
\arguments{
\item{p}{The probability at which to calculate the quantile.}

\item{mixdense}{An object of class \code{normalmix}.}
}
\description{
Quantile function for a mixture of normals.
}
\author{
David Gerard
}

