% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ash.R
\name{qval.from.lfdr}
\alias{qval.from.lfdr}
\title{Function to compute q values from local false discovery rates}
\usage{
qval.from.lfdr(lfdr)
}
\arguments{
\item{lfdr, }{a vector of local fdr estimates}
}
\value{
vector of q values
}
\description{
Computes q values from a vector of local fdr estimates
}
\details{
The q value for a given lfdr is an estimate of the (tail)
    False Discovery Rate for all findings with a smaller lfdr, and
    is found by the average of the lfdr for all more significant
    findings. See Storey (2003), Annals of Statistics, for
    definition of q value.
}
