% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\docType{data}
\name{sea}
\alias{sea}
\title{List of watery palettes, 8 or 5 values long.}
\format{
An object of class \code{list} of length 10.
}
\usage{
sea
}
\description{
List of watery palettes, 8 or 5 values long.
}
\keyword{datasets}
