% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_data.R
\name{set_data}
\alias{set_data}
\title{Takes raw data and sets up data object for use by ash}
\usage{
set_data(betahat, sebetahat, lik = NULL, alpha = 0)
}
\arguments{
\item{betahat}{vector of betahats}

\item{sebetahat}{vector of standard errors}

\item{lik}{a likelihood (see e.g., lik_normal())}

\item{alpha}{specifies value of alpha to use (model is for betahat/sebetahat^alpha | sebetahat)}
}
\value{
data object (list)
}
\description{
Takes raw data and sets up data object for use by ash
}
\details{
The data object stores both the data, and details of the model to be used for the data.
For example, in the generalized version of ash the cdf and pdf of the likelihood are
stored here.
}
