% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.ash}
\alias{summary.ash}
\title{Summary method for ash object}
\usage{
\method{summary}{ash}(a)
}
\description{
Print summary of fitted ash object
}
\details{
\code{\link{summary}} prints the fitted mixture, the fitted log likelihood with 10 digits and a flag to indicate convergence
}

