% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lik.R
\name{t_lik}
\alias{t_lik}
\title{Likelihood object for t error distribution}
\usage{
t_lik(df)
}
\arguments{
\item{df}{degree of freedom parameter of t distribution}
}
\description{
Creates a likelihood object for ash for use with t error distribution
}
\examples{
   z = rnorm(100) + rt(100,df=4) # simulate some data with t error
   ash(z,1,lik=t_lik(df=4))
}

