% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnormalmix.R
\name{tnormalmix}
\alias{tnormalmix}
\title{Constructor for tnormalmix class}
\usage{
tnormalmix(pi, mean, sd, a, b)
}
\arguments{
\item{pi}{vector of mixture proportions (length k say)}

\item{mean}{vector of means (length k)}

\item{sd}{vector of standard deviations (length k)}

\item{a}{vector of left truncation points of each component (length k)}

\item{b}{vector of right truncation points of each component (length k)}
}
\value{
an object of class tnormalmix
}
\description{
Creates an object of class tnormalmix (finite mixture
    of truncated univariate normals)
}
\details{
None
}
\examples{
tnormalmix(c(0.5,0.5),c(0,0),c(1,2),c(-10,0),c(0,10))

}
