% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unimix.R
\name{unimix}
\alias{unimix}
\title{Constructor for unimix class}
\usage{
unimix(pi, a, b)
}
\arguments{
\item{pi}{vector of mixture proportions}

\item{a}{vector of left hand ends of uniforms}

\item{b}{vector of right hand ends of uniforms}
}
\value{
an object of class unimix
}
\description{
Creates an object of class unimix (finite mixture of
    univariate uniforms)
}
\details{
None
}
\examples{
unimix(c(0.5,0.5),c(0,0),c(1,2))
}
