% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{vcdf_post}
\alias{vcdf_post}
\title{vcdf_post}
\usage{
vcdf_post(m, c, data)
}
\arguments{
\item{m}{mixture distribution with k components}

\item{c}{a numeric vector}

\item{data}{depends on context}
}
\value{
an n vector containing the cdf for beta_i at c
}
\description{
vectorized version of \code{\link{cdf_post}}
}
\examples{
beta = rnorm(100,0,1)
betahat= beta+rnorm(100,0,1)
sebetahat=rep(1,100)
ash.beta = ash(betahat,1,mixcompdist="normal")
c = vcdf_post(get_fitted_g(ash.beta),seq(-5,5,length=1000),data = set_data(betahat,sebetahat))
}
