% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mad_ash.R
\name{vlema}
\alias{vlema}
\title{Voom-limma-ebayes-mad-ash pipeline.}
\usage{
vlema(Y, X, cov_of_interest = ncol(X), likelihood = c("normal", "t"),
  ash_args = list(), include_intercept = TRUE)
}
\arguments{
\item{Y}{A matrix of numerics. These are the response variables
where each column has its own variance. In a gene expression
study, the rows are the individuals and the columns are the
genes.}

\item{X}{A matrix of numerics. The covariates of interest.}

\item{cov_of_interest}{A positive integer. The column number of the
covariate in X whose coefficients you want to apply ASH to.}

\item{likelihood}{Either \code{"normal"} or \code{"t"}. If
\code{likelihood = "t"}, then the user may provide the degrees
of freedom by including a \code{df} element in
\code{ash_args}. If \code{ash_args$df} is \code{NULL} then the
degrees of freedom will be the sample size minus the number of
covariates minus \code{k}.}

\item{ash_args}{A list of arguments to pass to ash. See
\code{\link{ash.workhorse}} for details.}

\item{include_intercept}{A logical. If \code{TRUE}, then it will
check \code{X} to see if it has an intercept term. If not, then
it will add an intercept term. If \code{FALSE}, then \code{X}
will be unchanged.}
}
\description{
Yuck.
}
\author{
David Gerard
}

