% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashutility.R
\name{as.data.frame.ash}
\alias{as.data.frame.ash}
\title{Create data from from ash object}
\usage{
\method{as.data.frame}{ash}(a, include_fdr = FALSE, ...)
}
\arguments{
\item{a}{the fitted ash object}

\item{include_fdr}{A logical, but never actually used.}

\item{...}{not used, included for consistency as an S3
generic/method.}
}
\description{
Creates data frame for easy plotting of results etc
}
\details{
Returns a data frame with named columnns
}

