% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ash.R
\name{ash}
\alias{ash}
\title{Main Adaptive Shrinkage function}
\usage{
ash(betahat, sebetahat, mixcompdist = c("uniform", "halfuniform", "normal"),
  df = NULL, ...)
}
\arguments{
\item{betahat}{a p vector of estimates}

\item{sebetahat}{a p vector of corresponding standard errors}

\item{mixcompdist}{distribution of components in mixture ( "uniform","halfuniform" or "normal"), the default is "uniform". If you believe your effects may be asymmetric, use "halfuniform". The use of "normal" is permitted only if df=NULL.}

\item{df}{appropriate degrees of freedom for (t) distribution of betahat/sebetahat, default is NULL which is actually treated as infinity (Gaussian)}
}
\value{
ash returns an object of \code{\link[base]{class}} "ash", a list with the following elements(or a  simplified list, if \eqn{onlylogLR=TRUE}, \eqn{minimaloutput=TRUE}   or \eqn{multiseqoutput=TRUE}) \cr
\item{fitted.g}{fitted mixture, either a normalmix or unimix}
\item{logLR}{log P(D|mle(pi)) - log P(D|null)}
\item{loglik}{log P(D|mle(pi))}
\item{PosteriorMean}{A vector consisting the posterior mean of beta from the mixture}
\item{PosteriorSD}{A vector consisting the corresponding posterior standard deviation}
\item{PositiveProb}{A vector of posterior probability that beta is positive}
\item{NegativeProb}{A vector of posterior probability that beta is negative}
\item{ZeroProb}{A vector of posterior probability that beta is zero}
\item{lfsr}{The local false sign rate}
\item{lfsra}{The local false sign rate(adjusted)}
\item{lfdr}{A vector of estimated local false discovery rate}
\item{qvalue}{A vector of q values}
\item{fit}{The fitted mixture object by \code{\link{mixEM}} or \code{\link{mixVBEM}} }
\item{lambda1}{multiplicative "inflation factor"}
\item{lambda2}{additive "inflation factor"}
\item{call}{a call in which all of the specified arguments are specified by their full names}
\item{data}{a list consisting the input betahat and sebetahat}
\item{excludeindex}{the vector of index of observations with 0 standard error; if none, then returns NULL}
\item{df}{the specified degrees of freedom for (t) distribution of betahat/sebetahat}
\item{model}{either "EE" or "ET", denoting whether exchangeable effects (EE) or exchangeable T stats (ET) has been used}
}
\description{
Takes vectors of estimates (betahat) and their standard errors (sebetahat), and applies
shrinkage to them, using Empirical Bayes methods, to compute shrunk estimates for beta.
}
\details{
This function is actually just a simple wrapper that passes its parameters to ash.workhorse which provides more documented options for advanced use. See readme for more details.
}
\examples{
beta = c(rep(0,100),rnorm(100))
sebetahat = abs(rnorm(200,0,1))
betahat = rnorm(200,beta,sebetahat)
beta.ash = ash(betahat, sebetahat)
summary(beta.ash)
plot(betahat,beta.ash$PosteriorMean,xlim=c(-4,4),ylim=c(-4,4))

CIMatrix=ashci(beta.ash,level=0.95)
print(CIMatrix)

#Illustrating the non-zero mode feature
betahat=betahat+5
beta.ash = ash(betahat, sebetahat)
plot(betahat,beta.ash$PosteriorMean)
summary(beta.ash)
betan.ash=ash(betahat, sebetahat,nonzeromode=TRUE)
plot(betahat, betan.ash$PosteriorMean)
summary(betan.ash)
}
\seealso{
\code{\link{ash.workhorse}} for complete specification of ash function

\code{\link{ashci}} for computation of credible intervals after getting the ash object return by \code{ash()}

\code{\link{ashm}} for Multi-model Adaptive Shrinkage function
}

