% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruv_ash.R
\name{ash_ruv}
\alias{ash_ruv}
\title{Perform RUV to estimate confounders, estimate scale using simple
MLE, run ash.}
\usage{
ash_ruv(Y, X, ctl, k = NULL, cov_of_interest = ncol(X), ash_args = list(),
  include_intercept = TRUE, gls = TRUE)
}
\arguments{
\item{Y}{A matrix of numerics. These are the response variables
where each column has its own variance. In a gene expression
study, the rows are the individuals and the columns are the
genes.}

\item{X}{A matrix of numerics. The covariates of interest.}

\item{ctl}{A vector of logicals of length \code{ncol(Y)}. If
position i is \code{TRUE} then position i is considered a
negative control.}

\item{k}{A non-negative integer.The number of unobserved
confounders. If not specified and the R package sva is
installed, then this function will estimate the number of
hidden confounders using th methods of Buja and Eyuboglu
(1992).}

\item{cov_of_interest}{A positive integer. The column number of
covariate in X whose coefficients you want to apply ASH to.}

\item{ash_args}{A list of arguments to pass to ash. See
\code{\link{ash.workhorse}} for details.}

\item{include_intercept}{A logical. If \code{TRUE}, then it will
check \code{X} to see if it has an intercept term. If not, then
it will add an intercept term. If \code{FALSE}, then \code{X}
will be unchanged.}

\item{gls}{A logical. Should we use generalized least squares
(\code{TRUE}) or ordinary least squares (\code{FALSE}) for
estimating the confounders? The OLS version is equivalent to
using RUV to estimate the confounders.}
}
\description{
Perform RUV to estimate confounders, estimate scale using simple
MLE, run ash.
}
\author{
David Gerard
}
\references{
Andreas Buja and Nermin Eyuboglu. Remarks on parallel
    analysis. Multivariate behavioral research, 27(4):509–540,
    1992.
}

