% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ashci}
\alias{ashci}
\title{Credible Interval Computation for the ash object}
\usage{
ashci(a, level = 0.95)
}
\arguments{
\item{a}{the fitted ash object}

\item{levels,}{the level for the credible interval, (default=0.95)}

\item{tol,}{the tolerance for convergence of log-likelihood.}
}
\value{
A matrix, with first column being the posterior mean, second and third column being the lower bound and upper bound for the credible interval.
}
\description{
Given the ash object return by the main function ash, this function computes the corresponding credible interval of the mixture model.
}
\details{
Uses default optimization function and perform component-wise credible interval computation. The computation cost is linear of the length of betahat.
}

