% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ashci}
\alias{ashci}
\title{Credible Interval Computation for the ash object}
\usage{
ashci(a, level = 0.95, tol = 1e-05, trace = FALSE)
}
\arguments{
\item{a}{the fitted ash object}

\item{tol}{the desired accuracy, default value is 1e-5.}

\item{trace}{a logical variable denoting whether some of the intermediate results of iterations should be displayed to the user. Default is FALSE.}

\item{levels}{the level for the credible interval, (default=0.95)}
}
\value{
A matrix, with first column being the posterior mean, second and third column being the lower bound and upper bound for the credible interval.
}
\description{
Given the ash object return by the main function ash, this function computes the corresponding credible interval of the mixture model.
}
\details{
Uses default optimization function and perform component-wise credible interval computation. The computation cost is linear of the length of betahat.
}
\examples{
beta = c(rep(0,20),rnorm(20))
sebetahat = abs(rnorm(40,0,1))
betahat = rnorm(40,beta,sebetahat)
beta.ash = ash(betahat, sebetahat)

CImatrix=ashci(beta.ash,level=0.95)
}

