% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autumn.R
\docType{data}
\name{autumn}
\alias{autumn}
\title{List of autumnal palettes, 8 or 5 values long.}
\format{
An object of class \code{list} of length 25.
}
\usage{
autumn
}
\description{
List of autumnal palettes, 8 or 5 values long.
}
\keyword{datasets}
