% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berry.R
\docType{data}
\name{berry}
\alias{berry}
\title{List of berry palettes, 8 or 5 values long.}
\format{
An object of class \code{list} of length 15.
}
\usage{
berry
}
\description{
List of berry palettes, 8 or 5 values long.
}
\keyword{datasets}
