% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lik.R
\name{binom_lik}
\alias{binom_lik}
\title{Likelihood object for Binomial error distribution}
\usage{
binom_lik(y, n)
}
\arguments{
\item{y}{Binomial observations}

\item{n}{Binomial number of trials}
}
\description{
Creates a likelihood object for ash for use with Binomial error distribution
}
\examples{
   p = rbeta(100,2,2) # prior mode: 0.5
   n = rpois(100,10)
   y = rbinom(100,n,p) # simulate Binomial observations
   ash(rep(0,length(y)),1,lik=binom_lik(y,n))
}

