% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ashutility.R
\name{calc_gloglik}
\alias{calc_gloglik}
\title{Compute loglikelihood for data usign the prior g (usually from an ash fit)}
\usage{
calc_gloglik(g, betahat, betahatsd, df, model = c("EE", "ES"))
}
\arguments{
\item{g}{the prior for effects or standardized effects}

\item{betahat}{the data}

\item{betahatsd}{the observed standard errors}

\item{df}{appropriate degrees of freedom for (t) distribution of betahat/sebetahat}

\item{model}{indicates whether you want the likelihood under the EE or ES model}
}
\description{
Return the log-likelihood of the data betahat, with standard errors betahatsd,
under the fitted distribution in the ash object.
}
\details{
See example in CompareBetahatvsZscoreAnalysis.rmd
}

