% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashutility.R
\name{calc_loglik}
\alias{calc_loglik}
\title{Compute loglikelihood for data from ash fit}
\usage{
calc_loglik(g, data)
}
\arguments{
\item{g}{the fitted g, or an ash object containing g}

\item{data}{a data object, see set_data}
}
\description{
Return the log-likelihood of the data for a given g() prior
}
