% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashutility.R
\name{calc_null_loglik}
\alias{calc_null_loglik}
\title{Compute loglikelihood for data under null that all beta are 0}
\usage{
calc_null_loglik(data)
}
\arguments{
\item{data}{a data object; see set_data}
}
\description{
Return the log-likelihood of the data betahat, with
    standard errors betahatsd, under the null that beta==0
}
