% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashutility.R
\name{calc_vlogLR}
\alias{calc_vlogLR}
\title{Compute vector of loglikelihood ratio for data from ash fit}
\usage{
calc_vlogLR(g, data)
}
\arguments{
\item{g}{the fitted g, or an ash object containing g}

\item{data}{a data object, see set_data}
}
\description{
Return the vector of log-likelihood ratios of the data
    betahat, with standard errors betahatsd, for a given g() prior
    on beta, or an ash object containing that, vs the null that g()
    is point mass on 0
}
