% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{comp_postmean}
\alias{comp_postmean}
\title{comp_postmean}
\usage{
comp_postmean(m, data)
}
\arguments{
\item{m}{mixture distribution with k components}

\item{data}{details depend on the model}
}
\description{
output posterior mean for beta for each component of
    prior mixture m,given data
}
