% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{comp_postmean2}
\alias{comp_postmean2}
\title{comp_postmean2}
\usage{
comp_postmean2(m, data)
}
\arguments{
\item{m}{mixture distribution with k components}

\item{data}{details depend on the model}
}
\description{
output posterior mean-squared value given prior
    mixture m and data
}
