% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{comp_postsd}
\alias{comp_postsd}
\title{comp_postsd}
\usage{
comp_postsd(m, betahat, sebetahat, v)
}
\arguments{
\item{m}{mixture distribution with k components}

\item{betahat}{an n vector of observations}

\item{sebetahat}{an n vector of standard errors}

\item{v}{degree of freedom of error distribution}
}
\description{
output posterior sd for beta for each component of
    prior mixture m,given observations betahat, sebetahat, df v
}
\examples{
beta = rnorm(100,0,1)
betahat= beta+rnorm(100,0,1)
ash.beta = ash(betahat,1,mixcompdist="normal")
comp_postmean(ash.beta$fitted.g,betahat,rep(1,100),NULL)
comp_postsd(ash.beta$fitted.g,betahat,rep(1,100),NULL)
comppostprob(ash.beta$fitted.g,betahat,rep(1,100),NULL)
}

