% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{comp_sd}
\alias{comp_sd}
\title{Generic function to extract the standard deviations of
    components of the mixture}
\usage{
comp_sd(m)
}
\arguments{
\item{m}{a mixture of k components generated by normalmix() or
unimix() or igmix()}
}
\value{
it returns a vector of standard deviations
}
\description{
Generic function to extract the standard deviations of
    components of the mixture
}
