% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mix.R
\name{compcdf_post.normalmix}
\alias{compcdf_post.normalmix}
\title{compcdf_post.normalmix}
\usage{
\method{compcdf_post}{normalmix}(m, c, betahat, sebetahat, v)
}
\arguments{
\item{m}{mixture distribution with k components}

\item{c}{a scalar}

\item{betahat}{an n vector of observations}

\item{sebetahat}{an n vector of standard errors}

\item{v}{degree of freedom of error distribution}
}
\value{
a k by n matrix
}
\description{
returns a k by n matrix of the posterior cdf evaluated at location c
}

