% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{compcdf_post}
\alias{compcdf_post}
\title{compcdf_post}
\usage{
compcdf_post(m, c, betahat, sebetahat, v)
}
\arguments{
\item{m}{mixture distribution with k components}

\item{c}{a scalar}

\item{betahat}{an n vector of observations}

\item{sebetahat}{an n vector of standard errors}

\item{v}{degree of freedom of error distribution}
}
\value{
a k by n matrix
}
\description{
evaluate cdf of posterior distribution of beta at c. m
    is the prior on beta, a mixture; c is location of evaluation
    assumption is betahat | beta ~ t_v(beta,sebetahat)
}
\examples{
beta = rnorm(100,0,1)
betahat= beta+rnorm(100,0,1)
sebetahat=rep(1,100)
ash.beta = ash(betahat,1,mixcompdist="normal")
compcdf_post(ash.beta$fitted.g,0,betahat,sebetahat,NULL)
}

