% Generated by roxygen2 (4.0.2): do not edit by hand
\name{compdens_conv}
\alias{compdens_conv}
\title{compdens_conv}
\usage{
compdens_conv(m, x, s, v, FUN = "+")
}
\arguments{
\item{m}{mixture distribution}

\item{x}{an n vector}

\item{s}{an n vector or integer}

\item{v}{degree of freedom of error distribution}
}
\value{
a k by n matrix of densities
}
\description{
compute the density of the components of the mixture m when convoluted with a normal with standard deviation s or a scaled (se) student.t with df v, the density is evaluated at x
}

