% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mix.R
\name{comppostprob.default}
\alias{comppostprob.default}
\title{comppostprob.default}
\usage{
\method{comppostprob}{default}(m, x, s, v)
}
\arguments{
\item{m}{mixture distribution}

\item{x}{an n vector}

\item{s}{an n vector or integer}

\item{v}{degree of freedom of error distribution}
}
\description{
comppostprob.default
}

