% Generated by roxygen2 (4.0.2): do not edit by hand
\name{compute_lfsr}
\alias{compute_lfsr}
\title{Faster version of function ash
#
#}
\usage{
compute_lfsr(NegativeProb, ZeroProb)
}
\arguments{
\item{betahat,}{a p vector of estimates
#}

\item{sebetahat,}{a p vector of corresponding standard errors
#}

\item{nullcheck:}{whether to check that any fitted model exceeds the "null" likelihood in which all weight is on the first component
#}

\item{randomstart:}{logical, indicating whether to initialize EM randomly. If FALSE, then initializes to prior mean (for EM algorithm) or prior (for VBEM)
#}

\item{pointmass:}{logical, indicating whether to use a point mass at zero as one of components for a mixture distribution
#}

\item{onlylogLR:}{logical, indicating whether to use this function to get logLR. Skip posterior prob, posterior mean, lfdr...
#}

\item{prior:}{string, or numeric vector indicating Dirichlet prior on mixture proportions (defaults to "uniform", or 1,1...,1; also can be "nullbiased" 1,1/k-1,...,1/k-1 to put more weight on first component)
#}

\item{mixsd:}{vector of sds for underlying mixture components
#}

\item{VB:}{whether to use Variational Bayes to estimate mixture proportions (instead of EM to find MAP estimate)
#}

\item{gridmult:}{the multiplier by which the default grid values for mixsd differ by one another. (Smaller values produce finer grids)
#}

\item{g:}{the prior distribution for beta (usually estimated from the data; this is used primarily in simulated data to do computations with the "true" g)
#}

\item{cxx:}{flag to indicate whether to use the c++ (Rcpp) version
#
#}
}
\value{
a list with elements fitted.g is fitted mixture
#logLR : logP(D|mle(pi)) - logP(D|null)
#
#
}
\description{
This function has similar functionality as ash, but only returns some of the outputs.
#
#
}
\details{
#
}

