% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ash.R
\name{compute_lfsr}
\alias{compute_lfsr}
\title{Function to compute the local false sign rate}
\usage{
compute_lfsr(NegativeProb, ZeroProb)
}
\arguments{
\item{NegativeProb}{A vector of posterior probability that beta is
negative.}

\item{ZeroProb}{A vector of posterior probability that beta is
zero.}
}
\value{
The local false sign rate.
}
\description{
Function to compute the local false sign rate
}
