% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{dens_conv}
\alias{dens_conv}
\title{dens_conv}
\usage{
dens_conv(m, data)
}
\arguments{
\item{m}{mixture distribution with k components}

\item{data}{details depend on the model}
}
\description{
compute density of mixture m convoluted with normal of
    sd (s) or student t with df v at locations x
}
