% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\docType{data}
\name{earth}
\alias{earth}
\title{List of earth-tone palettes, 8 or 5 values long.}
\format{
An object of class \code{list} of length 5.
}
\usage{
earth
}
\description{
List of earth-tone palettes, 8 or 5 values long.
}
\keyword{datasets}
