% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ash.R
\name{estimate_mixprop}
\alias{estimate_mixprop}
\title{Estimate mixture proportions of sigmaa by EM algorithm}
\usage{
estimate_mixprop(data, g, prior, optmethod = c("mixEM", "mixVBEM",
  "cxxMixSquarem", "mixIP"), control = list())
}
\arguments{
\item{data}{list to be passed to log_compdens_conv; details depend on model}

\item{g}{the prior distribution for beta (usually estimated from
the data}

\item{prior}{string, or numeric vector indicating Dirichlet prior
on mixture proportions (defaults to "uniform", or (1,1...,1);
also can be "nullbiased" (nullweight,1,...,1) to put more
weight on first component)}

\item{optmethod}{name of function to use to do optimization}

\item{control}{A list of control parameters for the SQUAREM
algorithm, default value is set to be control.default=list(K =
1, method=3, square=TRUE, step.min0=1, step.max0=1, mstep=4,
kr=1, objfn.inc=1,tol=1.e-07, maxiter=5000, trace=FALSE).}
}
\value{
A list, including the final loglikelihood, the null loglikelihood, a n by k likelihoodmatrix with (j,k)th element equal to \eqn{f_k(x_j)},and a flag to indicate convergence.
}
\description{
Estimate mixture proportions of sigmaa by EM algorithm
}

