% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashutility.R
\name{get_density}
\alias{get_density}
\title{Density method for ash object}
\usage{
get_density(a, x)
}
\arguments{
\item{a}{the fitted ash object}

\item{x}{the vector of locations at which density is to be computed}
}
\description{
Return the density of the underlying fitted distribution
}
\details{
None
}
