% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashutility.R
\name{get_loglik}
\alias{get_loglik}
\title{Get fitted loglikelihood for ash object}
\usage{
get_loglik(a)
}
\arguments{
\item{a}{the fitted ash object}
}
\description{
Return the log-likelihood of the data under the fitted
    distribution
}
\details{
None
}

