\name{loglik.ash}
\alias{loglik.ash}
\title{Compute loglikelihood for data from ash fit}
\usage{
  loglik.ash(a, betahat, betahatsd)
}
\arguments{
  \item{a}{the fitted ash object}

  \item{betahat}{the data}

  \item{betahatsd}{the observed standard errors}
}
\description{
  Return the log-likelihood of the data betahat, with
  standard errors betahatsd, under the fitted distribution
  in the ash object
}
\details{
  None
}

