% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{loglik_conv.default}
\alias{loglik_conv.default}
\title{loglik_conv.default}
\usage{
\method{loglik_conv}{default}(m, betahat, betahatsd, v, FUN = "+")
}
\arguments{
\item{m}{mixture distribution with k components}

\item{betahat}{an n vector of observations}

\item{betahatsd}{an n vector of standard errors}

\item{v}{degree of freedom of error distribution}

\item{FUN}{default is "+"}
}
\description{
The default version of \code{\link{loglik_conv}}.
}

