% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{loglik_conv.default}
\alias{loglik_conv.default}
\title{loglik_conv.default}
\usage{
\method{loglik_conv}{default}(m, data)
}
\arguments{
\item{m}{mixture distribution with k components}

\item{data}{data whose details depend on model}
}
\description{
The default version of \code{\link{loglik_conv}}.
}
