% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{loglik_conv}
\alias{loglik_conv}
\title{loglik_conv}
\usage{
loglik_conv(m, betahat, betahatsd, v, FUN = "+")
}
\arguments{
\item{m}{mixture distribution with k components}

\item{betahat}{an n vector of observations}

\item{betahatsd}{an n vector of standard errors}

\item{v}{degree of freedom of error distribution}

\item{FUN}{default is "+"}
}
\description{
find log likelihood of data in betahat, when the
    mixture m is convolved with a normal with sd betahatsd
}

