% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{loglik_conv}
\alias{loglik_conv}
\title{loglik_conv}
\usage{
loglik_conv(m, data)
}
\arguments{
\item{m}{mixture distribution with k components}

\item{data}{details depend on the model}
}
\description{
find log likelihood of data using convolution of mixture with error distribution
}
