% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{mixmean2}
\alias{mixmean2}
\title{Generic function of calculating the overall second moment of the
    mixture}
\usage{
mixmean2(m)
}
\arguments{
\item{m}{a mixture of k components generated by normalmix() or
unimix() or igmix()}
}
\value{
it returns scalar
}
\description{
Generic function of calculating the overall second moment of the
    mixture
}
