% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunct.R
\name{my_etrunct}
\alias{my_etrunct}
\title{my_etrunct}
\usage{
my_etrunct(a, b, df)
}
\arguments{
\item{a}{left limit of distribution}

\item{b}{right limit of distribution}

\item{df}{degree of freedom of error distribution}
}
\description{
Compute second moment of the truncated t. Uses results from O'Hagan, Biometrika, 1973
}
