% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonzero_mode.R
\name{nonzeromodeEM}
\alias{nonzeromodeEM}
\title{Estimate unimodal nonzero mean of a mixture model by EM algorithm}
\usage{
nonzeromodeEM(betahat, sebetahat, mixsd, mixcompdist, df = NULL,
  pi_init = NULL, control = list())
}
\arguments{
\item{betahat}{a p vector of estimates.}

\item{sebetahat}{a p vector of corresponding standard errors.}

\item{mixsd}{vector of sds for underlying mixture components.}

\item{mixcompdist}{distribution of components in mixture ( "uniform","halfuniform" or "normal").}

\item{df}{appropriate degrees of freedom for (t) distribution of betahat/sebetahat, default is NULL(Gaussian).}

\item{pi_init}{the initial value of \eqn{\pi} to use. If not specified defaults to (1/k,...,1/k).}

\item{control}{A list of control parameters for the SQUAREM algorithm, default value is set to be   control.default=list(K = 1, method=3, square=TRUE, step.min0=1, step.max0=1, mstep=4, kr=1, objfn.inc=1,tol=1.e-07, maxiter=5000, trace=FALSE).}
}
\value{
A list, including the estimates (\eqn{\mu}) and (\eqn{\pi}), the log likelihood for each iteration (NQ)
and a flag to indicate convergence.
}
\description{
Given the data, standard error of the data and standard deviations of the Gaussian mixture model, estimate the mean of a unimodal Gaussian mixture by an EM algorithm.
}
\details{
Fits a k component mixture model \deqn{f(x|\pi) = \sum_k \pi_k f_k(x)} to independent
and identically distributed data \eqn{x_1,\dots,x_n}.
Estimates unimodal mean \eqn{\mu} by EM algorithm. Uses the SQUAREM package to accelerate convergence of EM. Used by the ash main function; there is no need for a user to call this
function separately, but it is exported for convenience.
}

