% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashutility.R
\name{plot.ash}
\alias{plot.ash}
\title{Plot method for ash object}
\usage{
\method{plot}{ash}(x, ..., xmin, xmax)
}
\arguments{
\item{x}{the fitted ash object}

\item{...}{Arguments to be passed to methods,such as graphical parameters (see \code{\link[graphics]{plot}})}

\item{xmin}{xlim lower range, default is the lowest value of betahat}

\item{xmax}{xlim upper range, default is the highest value of betahat}
}
\description{
Plot the cdf of the underlying fitted distribution
}
\details{
None
}
