% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lik.R
\name{pois_lik}
\alias{pois_lik}
\title{Likelihood object for Poisson error distribution}
\usage{
pois_lik(y, link = c("identity", "log"))
}
\arguments{
\item{y}{Poisson observations}

\item{link}{Link function. The "identity" link directly puts unimodal prior on Poisson
intensities lambda, and "log" link puts unimodal prior on log(lambda).}
}
\description{
Creates a likelihood object for ash for use with Poisson error distribution
}
\examples{
   beta = c(rnorm(100,50,5)) # prior mode: 50
   x = rpois(100,beta) # simulate Poisson observations
   ash(rep(0,length(x)),1,lik=pois_lik(x))
}

