% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mix.R
\name{postmean2_mixlik}
\alias{postmean2_mixlik}
\title{postmean2_mixlik}
\usage{
postmean2_mixlik(m, betahat, sebetahat, v, pilik)
}
\arguments{
\item{m}{mixture distribution}

\item{betahat}{the data}

\item{sebetahat}{the observed standard errors}

\item{v}{degree of freedom of error distribution}

\item{pilik}{mixture proportion}
}
\description{
output posterior mean-squared value for beta for prior mixture m,given observations betahat, sebetahat, df v, from l-components mixture likelihood with mixture proportion pilik
}

