% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{postsd}
\alias{postsd}
\title{postsd}
\usage{
postsd(m, data)
}
\arguments{
\item{m}{mixture distribution with k components}

\item{data}{details depend on the model}
}
\description{
output posterior sd given prior mixture m and data
}
