% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashutility.R
\name{summary.ash}
\alias{summary.ash}
\title{Summary method for ash object}
\usage{
\method{summary}{ash}(object, ...)
}
\arguments{
\item{object}{the fitted ash object}

\item{...}{not used, included for consistency as an S3
generic/method.}
}
\description{
Print summary of fitted ash object
}
\details{
\code{\link{summary}} prints the fitted mixture, the
    fitted log likelihood with 10 digits and a flag to indicate
    convergence
}
